Move utils to a utils folder
