+++ src/ftrace.js
@@ -1,6 +1,6 @@
 "use strict";
 
-const parserHelpers = require('./parserHelpers')
+const parserHelpers = require('./utils/parserHelpers')
 const fs = require('fs')
 const parser = require('solidity-parser-antlr')
 const colors = require('colors')
